add : add feature ratings

+++ main server/src/main/java/ru/practicum/explore/model/user/User.java
@@ -19,5 +19,5 @@ public class User {
     private String name;
     @Column(unique = true, nullable = false)
     private String email;
-    private Float rate;
+    private Double rate;
 }
